safari-600.7-branch/LayoutTests/js/dom/Promise-resolve-with-then-fulfill.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;
var value = {
  then: function(fulfillCallback, rejectCallback) {
    testPassed('value.then is called.');
    window.thisValue = this;
    shouldBe('thisValue', 'value');
    fulfillCallback('hello');
  }
};
var promise = new Promise(function(resolve) { resolve(value); });

promise.then(function(result) {
  testPassed('fulfilled');
  window.result = result;
  shouldBeEqualToString('result', 'hello');
  finishJSTest();
}, function() {
  testFailed('rejected');
  finishJSTest();
});

debug('The promise is not fulfilled now.');

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
